+++ CHANGES.rst
@@ -14,6 +14,7 @@ Changelog
 
 **Changed**
 
+- #1428 Publish verified partitions
 - #1429 Add2: Do not set template values on already filled fields
 - #1427 Improved performance of Sample header table rendering
 - #1417 Cache allowed transitions for analyses on the request